Tela de login adicionada. Foram criados 3 pacotes diferentes e os arquivos antigos foram movidos para model.versaoAnterior
