+++ .gitignore
@@ -37,6 +37,7 @@ target/
 .project
 
 # NetBeans specific
+VoceAluga/bin #adicionado para remover fxmls duplicados
 nbproject/private/
 build/
 nbbuild/
