+++ VoceAluga/src/model/Gerente.java
@@ -1,0 +1,6 @@
+package model;
+
+public class Gerente extends Funcionario{
+	
+	
+}
