+++ VoceAluga/src/model/versaoAnterior/Main.java
@@ -1,4 +1,4 @@
-package com.company;
+package model.versaoAnterior;
 
 
 import java.util.LinkedHashMap;
