+++ VoceAluga/src/model/versaoAnterior/cadastroCliente.java
@@ -1,0 +1,5 @@
+package model.versaoAnterior;
+
+public class cadastroCliente {
+
+}
