+++ VoceAluga/src/view/Login.fxml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.PasswordField?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.Pane?>
+
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.LoginController">
+   <children>
+      <TextField fx:id="user" layoutX="355.0" layoutY="231.0" />
+      <PasswordField fx:id="password" layoutX="355.0" layoutY="280.0" />
+      <Button fx:id="entrar" layoutX="371.0" layoutY="346.0" mnemonicParsing="false" onMouseClicked="#processarLoginBotao" text="Entrar" />
+      <Label fx:id="label" layoutX="275.0" layoutY="236.0" text="Usuário:" />
+      <Label layoutX="275.0" layoutY="285.0" text="Senha:" />
+   </children>
+</Pane>
