--- VoceAluga_Grupo9/src/com/company/Apolice.java
@@ -1,12 +1,0 @@
-package com.company;
-
-public enum Apolice {
-    APOLICE_1(1),APOLICE_2(2),APOLICE_3(3),APOLICE_4(4),APOLICE_5(5),APOLICE_6(6),
-    APOLICE_7(7),APOLICE_8(8),APOLICE_9(9),APOLICE_10(10),APOLICE_11(11),APOLICE_12(12);
-
-    private int numeroDaApolice;
-
-    Apolice(int numero){
-        this.numeroDaApolice = numero;
-    }
-}
