--- VoceAluga_Grupo9/src/com/company/CadastroCliente.java
@@ -1,5 +1,0 @@
-package com.company;
-
-public final class CadastroCliente {
-
-}
