--- VoceAluga_Grupo9/src/com/company/Contrato.java
@@ -1,27 +1,0 @@
-package com.company;
-
-public class Contrato {
-    private Cliente assinante;
-    private boolean validez;
-
-    public Contrato(Cliente cliente){
-        this.assinante = cliente;
-    }
-
-    public void setValidez() {
-        this.validez = true;
-    }
-
-    /*
-    public void quebraDeContrato(){
-        this.validez = false;
-    }
-    */
-    public String getAssinante() {
-        return this.assinante.getCpf();
-    }
-
-    public void mostrarAssinante() {
-        System.out.printf("%s\n", this.assinante.getCpfFormatado());
-    }
-}
