--- VoceAluga_Grupo9/src/com/company/Pais.java
@@ -1,17 +1,0 @@
-package com.company;
-
-public enum Pais {
-    AFRICA_DO_SUL("ÁFRICA DO SUL"),ALEMANHA("ALEMANHA"),ARGENTINA("ARGENTINA"),
-    AUSTRALIA("AUSTRÁLIA"),BELGICA("BÉLGICA"),BOLIVIA("BOLÍVIA"),
-    BRASIL("BRASIL"),BULGARIA("BULGÁRIA"),CANADA("CANADÁ"),
-    CHILE("CHILE"),CHINA("CHINA"),COLOMBIA("COLÔMBIA"),
-    ESPANHA("ESPANHA"),ESTADOS_UNIDOS("ESTADOS UNIDOS"),FRANCA("FRANÇA"),
-    INDIA("ÍNDIA"),JAPAO("JAPÃO"),MARROCOS("MARROCOS"),
-    MEXICO("MÉXICO"),PORTUGAL("PORTUGAL"),RUSSIA("RÚSSIA");
-
-    public final String nomePais;
-
-    Pais(String nomePais){
-        this.nomePais = nomePais;
-    }
-}
